+++ ui/apps/platform/cypress/helpers/clusters.js
@@ -123,3 +123,26 @@ export function visitClusterByNameWithFixtureMetadataDatetime(
         cy.get(selectors.clusterSidePanelHeading).contains(clusterName);
     });
 }
+
+export function visitDashboardWithNoClusters() {
+    cy.intercept('POST', api.graphql('summary_counts'), {
+        body: {
+            data: {
+                clusterCount: 0,
+                nodeCount: 3,
+                violationCount: 20,
+                deploymentCount: 35,
+                imageCount: 31,
+                secretCount: 15,
+            },
+        },
+    }).as('summary_counts');
+    cy.intercept('GET', api.clusters.list, {
+        clusters: [],
+    }).as('clusters');
+
+    // visitMainDashboard(); // with a count of 0 clusters, app should redirect to the clusters pages
+    cy.visit('/main/dashboard'); // with a count of 0 clusters, app should redirect to the clusters pages
+
+    cy.wait(['@summary_counts', '@clusters']);
+}